Generalize type signature MethodSpec#addExceptions
This is better because it allows me to pass something like FluentIterable<ClassName>, without having to convert it into a list and then double-casting it. Please consider auditing your API for other things like this.

This change should be trivial enough that it won't any of that legal mumbo jumbo.